backend/requirements.txt
django==3.2.18
djangorestframework==3.14
djangorestframework-simplejwt==4.4.0
django-cors-headers==3.10.1